refactor: 공용이 아닌 컴포넌트 위치 수정
