+++ app/.internal/components/InternalEditor.tsx
@@ -2,7 +2,7 @@
 
 import { FormProvider, useForm } from 'react-hook-form';
 
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import Form from '@/components/form/Form';
 import { useRouter } from '@/i18n/routing';
 import { errorToStr } from '@/utils/error';
