+++ app/[locale]/about/components/AboutEditor.tsx
@@ -9,11 +9,10 @@ import {
   putHistoryAction,
   putOverviewAction,
 } from '@/actions/about';
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
+import Form from '@/components/form/Form';
 import LanguagePicker from '@/components/form/LanguagePicker';
 import { PostEditorFile, PostEditorImage } from '@/components/form/types';
-import Form from '@/components/form/Form';
-import HTMLEditor from '@/components/form/html/HTMLEditor';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { useRouter } from '@/i18n/routing';
 import { AboutContent } from '@/types/about';
