+++ app/[locale]/about/future-careers/description/edit/CareerDescriptionEditor.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { putCareerDescriptionAction } from '@/actions/about';
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import LanguagePicker from '@/components/form/LanguagePicker';
 import Form from '@/components/form/Form';
 import HTMLEditor from '@/components/form/html/HTMLEditor';
