+++ app/[locale]/academics/components/courses/AddCourseModal.tsx
@@ -1,11 +1,11 @@
 import { FormProvider, useForm } from 'react-hook-form';
 
 import { postCourseAction } from '@/actions/academics';
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import Form from '@/components/form/Form';
 import ModalFrame from '@/components/modal/ModalFrame';
 import { CLASSIFICATION, Course, GRADE, StudentType } from '@/types/academics';
-import { getKeys } from '@/types/object';
+import { getKeys } from '@/utils/object';
 import { errorToStr } from '@/utils/error';
 import { handleServerAction } from '@/utils/serverActionError';
 import { errorToast, successToast } from '@/utils/toast';
