+++ app/[locale]/academics/components/timeline/TimelineEditor.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from 'next/router';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import { isUploadedFile, PostEditorFile } from '@/components/form/types';
 import Form from '@/components/form/Form';
 import HTMLEditor from '@/components/form/html/HTMLEditor';
