+++ app/[locale]/community/components/PostDeleteButton.tsx
@@ -5,10 +5,9 @@ import { useTransition } from 'react';
 import { deleteNewsAction } from '@/actions/news';
 import { deleteNoticeAction } from '@/actions/notice';
 import { deleteSeminarAction } from '@/actions/seminar';
+import AlertModal from '@/components/modal/AlertModal';
 import useModal from '@/utils/hooks/useModal';
-
-import { errorToast, successToast } from '../../utils/toast';
-import AlertModal from '../modal/AlertModal';
+import { errorToast, successToast } from '@/utils/toast';
 
 export default function PostDeleteButton({ postType, id }: { postType: string; id: string }) {
   const { openModal } = useModal();
