+++ app/[locale]/community/faculty-recruitment/edit/FacultyRecruitmentEditPageContent.tsx
@@ -3,7 +3,7 @@
 import { FormProvider, useForm } from 'react-hook-form';
 
 import { putFacultyRecruitmentAction } from '@/actions/recruitment';
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import { PostEditorImage } from '@/components/form/types';
 import Form from '@/components/form/Form';
 import HTMLEditor from '@/components/form/html/HTMLEditor';
