+++ app/[locale]/community/news/components/NewsEditor.tsx
@@ -1,6 +1,6 @@
 import { FormProvider, useForm } from 'react-hook-form';
 
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import { PostEditorFile, PostEditorImage } from '@/components/form/types';
 import Form from '@/components/form/Form';
 import HTMLEditor from '@/components/form/html/HTMLEditor';
