+++ app/[locale]/community/news/components/NewsRow.tsx
@@ -1,6 +1,6 @@
+import PaginatedLink from '@/app/[locale]/community/components/PaginatedLink';
 import ImageWithFallback from '@/components/common/ImageWithFallback';
 import Tags from '@/components/common/Tags';
-import PaginatedLink from '@/components/post/PaginatedLink';
 import { getPath } from '@/utils/page';
 import { news } from '@/utils/segmentNode';
