+++ app/[locale]/community/notice/components/NoticeListRow.tsx
@@ -1,4 +1,4 @@
-import PaginatedLink from '@/components/post/PaginatedLink';
+import PaginatedLink from '@/app/[locale]/community/components/PaginatedLink';
 import CheckboxOrange from '@/public/image/checkbox_orange.svg';
 import ClipIcon from '@/public/image/clip_icon.svg';
 import LockIcon from '@/public/image/lock_icon.svg';
