+++ app/[locale]/community/seminar/[id]/SeminarViewer.tsx
@@ -1,12 +1,12 @@
 import { getTranslations } from 'next-intl/server';
 import { ReactNode } from 'react';
 
+import PostFooter from '@/app/[locale]/community/components/PostFooter';
 import Attachments from '@/components/common/Attachments';
 import ImageWithFallback from '@/components/common/ImageWithFallback';
 import { StraightNode } from '@/components/common/Nodes';
 import HTMLViewer from '@/components/form/html/HTMLViewer';
 import { PAGE_PADDING_BOTTOM_PX } from '@/components/layout/pageLayout/PageLayout';
-import PostFooter from '@/components/post/PostFooter';
 import { Link } from '@/i18n/routing';
 import { Seminar } from '@/types/seminar';
