+++ app/[locale]/community/seminar/[id]/page.tsx
@@ -2,7 +2,7 @@ import { Suspense } from 'react';
 
 import { getSeminarPost } from '@/apis/v1/seminar/[id]';
 import InvalidIDFallback from '@/components/common/InvalidIDFallback';
-import PostFallback from '@/components/layout/fallback/PostFallback';
+import PostFallback from '@/app/[locale]/community/components/PostFallback';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { PostSearchQueryParams } from '@/types/post';
 import { getMetadata } from '@/utils/metadata';
