+++ app/[locale]/components/GraphicSection.tsx
