+++ app/[locale]/components/NewsCard.tsx
