+++ app/[locale]/components/NewsCarousel.tsx
