+++ app/[locale]/components/NewsCarouselMobile.tsx
