+++ app/[locale]/components/NewsSection.tsx
