+++ app/[locale]/components/NoticeSection.tsx
