+++ app/[locale]/components/animateScrollTo.ts
