+++ app/[locale]/components/constants.ts
