+++ app/[locale]/components/useCarousel.ts
