+++ app/[locale]/components/useCarouselLayout.tsx
