+++ app/[locale]/page.tsx
@@ -3,9 +3,9 @@ import { getTranslations } from 'next-intl/server';
 
 import { getMain } from '@/apis/v1';
 import Header from '@/components/layout/header/Header';
-import GraphicSection from '@/components/main/GraphicSection';
-import NewsSection from '@/components/main/NewsSection';
-import NoticeSection from '@/components/main/NoticeSection';
+import GraphicSection from '@/app/[locale]/components/GraphicSection';
+import NewsSection from '@/app/[locale]/components/NewsSection';
+import NoticeSection from '@/app/[locale]/components/NoticeSection';
 import { Link } from '@/i18n/routing';
 import { MainImportant } from '@/types/main';
 import { getPath } from '@/utils/page';
