+++ app/[locale]/research/centers/components/ResearchCenterEditor.tsx
@@ -8,7 +8,7 @@ import { FormProvider, useForm } from 'react-hook-form';
 import LanguagePicker from '@/components/form/LanguagePicker';
 import { Language } from '@/types/language';
 import { useState } from 'react';
-import Fieldset from '@/components/form/fieldset';
+import Fieldset from '@/components/form/Fieldset';
 import Form from '@/components/form/Form';
 import HTMLEditor from '@/components/form/html/HTMLEditor';
 import { PostEditorImage } from '@/components/form/types';
