+++ app/[locale]/research/labs/components/ResearchLabEditor.tsx
@@ -10,7 +10,7 @@ import { ResearchGroup } from '@/types/research';
 import { getPath } from '@/utils/page';
 import { researchLabs } from '@/utils/segmentNode';
 
-import Fieldset from '../../../../../components/form/fieldset';
+import Fieldset from '../../../../../components/form/Fieldset';
 import LanguagePicker from '../../../../../components/form/LanguagePicker';
 import { PostEditorFile } from '../../../../../components/form/types';
 import Form from '../../../../../components/form/Form';
