+++ components/layout/footer/constants.ts
