--- types/admissions.ts
@@ -1,3 +1,0 @@
-export interface Admission {
-  description: string;
-}
