--- utils/number.ts
@@ -1,2 +1,0 @@
-const NUMBER_REGEX = /^\d*$/;
-export const isNumber = (value: string) => NUMBER_REGEX.test(value);
